Extract codespaces interface for LiveshareSession
