+++ pkg/liveshare/client.go
@@ -20,7 +20,6 @@ import (
 	"time"
 
 	"github.com/opentracing/opentracing-go"
-	"golang.org/x/crypto/ssh"
 )
 
 type logger interface {
@@ -135,42 +134,4 @@ type joinWorkspaceArgs struct {
 
 type joinWorkspaceResult struct {
 	SessionNumber int `json:"sessionNumber"`
-}
-
-// A channelID is an identifier for an exposed port on a remote
-// container that may be used to open an SSH channel to it.
-type channelID struct {
-	name, condition string
-}
-
-func (s *Session) openStreamingChannel(ctx context.Context, id channelID) (ssh.Channel, error) {
-	type getStreamArgs struct {
-		StreamName string `json:"streamName"`
-		Condition  string `json:"condition"`
-	}
-	args := getStreamArgs{
-		StreamName: id.name,
-		Condition:  id.condition,
-	}
-	var streamID string
-	if err := s.rpc.do(ctx, "streamManager.getStream", args, &streamID); err != nil {
-		return nil, fmt.Errorf("error getting stream id: %w", err)
-	}
-
-	span, ctx := opentracing.StartSpanFromContext(ctx, "Session.OpenChannel+SendRequest")
-	defer span.Finish()
-	_ = ctx // ctx is not currently used
-
-	channel, reqs, err := s.ssh.conn.OpenChannel("session", nil)
-	if err != nil {
-		return nil, fmt.Errorf("error opening ssh channel for transport: %w", err)
-	}
-	go ssh.DiscardRequests(reqs)
-
-	requestType := fmt.Sprintf("stream-transport-%s", streamID)
-	if _, err = channel.SendRequest(requestType, true, nil); err != nil {
-		return nil, fmt.Errorf("error sending channel request: %w", err)
-	}
-
-	return channel, nil
 }
