+++ pkg/liveshare/session_test.go
@@ -103,7 +103,7 @@ func TestServerStartSharing(t *testing.T) {
 
 	done := make(chan error)
 	go func() {
-		streamID, err := session.startSharing(ctx, serverProtocol, serverPort)
+		streamID, err := session.StartSharing(ctx, serverProtocol, serverPort)
 		if err != nil {
 			done <- fmt.Errorf("error sharing server: %w", err)
 		}
@@ -247,10 +247,10 @@ func TestInvalidHostKey(t *testing.T) {
 func TestKeepAliveNonBlocking(t *testing.T) {
 	session := &Session{keepAliveReason: make(chan string, 1)}
 	for i := 0; i < 2; i++ {
-		session.keepAlive("io")
+		session.KeepAlive("io")
 	}
 
-	// if keepAlive blocks, we'll never reach this and timeout the test
+	// if KeepAlive blocks, we'll never reach this and timeout the test
 	// timing out
 }
 
@@ -367,10 +367,10 @@ func TestSessionHeartbeat(t *testing.T) {
 
 	go session.heartbeat(ctx, 50*time.Millisecond)
 	go func() {
-		session.keepAlive("input")
+		session.KeepAlive("input")
 		wg.Wait()
 		wg.Add(1)
-		session.keepAlive("input")
+		session.KeepAlive("input")
 		wg.Wait()
 		done <- struct{}{}
 	}()
@@ -380,7 +380,7 @@ func TestSessionHeartbeat(t *testing.T) {
 		t.Errorf("error from server: %v", err)
 	case <-done:
 		activityCount := strings.Count(logger.String(), "input")
-		// by design keepAlive can drop requests, and therefore there is zero guarantee
+		// by design KeepAlive can drop requests, and therefore there is zero guarantee
 		// that we actually get two requests if the network happened to be slow (rarely)
 		// during testing.
 		if activityCount != 1 && activityCount != 2 {
